% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{ldap_query}
\alias{ldap_query}
\title{Generic LDAP query}
\usage{
ldap_query(
  hostname,
  base_dn = "",
  attributes = "",
  scope = c("base", "one", "sub"),
  filter = "(objectClass=*)",
  user,
  pass,
  protocol = c("ldaps", "ldap"),
  port = c(636, 389)
)
}
\arguments{
\item{hostname}{name (or ip address) of the LDAP server}

\item{base_dn}{distinguished name (DN) of an entry in the directory. This
DN identifies the entry that is the starting point of the search. If no base
DN is specified, the search starts at the root of the directory tree.}

\item{attributes}{The attributes to be returned. To specify more than one
attribute, use commas to separate the attributes (for example,
"cn,mail,telephoneNumber"). If no attributes are specified in the URL,
all attributes are returned.}

\item{scope}{The scope of the search, which can be one of these values:
\enumerate{
\item base
\itemize{
\item retrieves information about the distinguished name (base_dn) specified
in the URL only.
}
\item one
\itemize{
\item retrieves information about entries one level below the distinguished
name (base_dn) specified in the URL. The base entry is not included in
this scope.
}
\item sub
\itemize{
\item retrieves information about entries at all levels below the
distinguished name (base_dn) specified in the URL. The base entry is
included in this scope.
}
}}

\item{filter}{Search filter to apply to entries within the specified scope of
the search.}

\item{user}{username}

\item{pass}{password}

\item{protocol}{connection protocol to use}

\item{port}{override the default port number (LDAP = 389, LDAPS = 636)}
}
\value{
curl response object
}
\description{
Perform generic LDAP queries
}
\examples{
ldap_query(
  hostname = "ldap.forumsys.com",
  base_dn = "dc=example,dc=com",
  scope = "sub",
  user = "cn=read-only-admin,dc=example,dc=com",
  pass = "password",
  protocol = "ldap",
  port = 389
) \%>\%
  content_to_tibble()

}
